requirements.txt
discord.py[voice]~=2.3.2
PyNaCl~=1.5.0
python-dotenv~=1.0.0
git+https://github.com/ytdl-org/youtube-dl.git@master#egg=youtube_dl
flask~=3.0.0
async-timeout~=4.0.3